Neural module unit test refactor - middle of core unit
